test: mutation changeGameStatus
