+++ config/constants.js
@@ -3,5 +3,5 @@ module.exports = {
         GUESSING_MODE: 60000,
         DRAWING_MODE: 90000
     },
-    DEFAULT_EXPIRATION: process.env.NODE_ENV === 'production' ? 900 : 1800
+    DEFAULT_MODEL_EXPIRATION: process.env.NODE_ENV === 'production' ? 900 : 1800
 }
